Move passphraseInput component to its own folder
