+++ src/components/passphraseInput/index.js
